Refactor containers so we can avoid circular dependencies
