+++ screens/BlockScreen/index.js
@@ -1,19 +1,19 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import BlockContents from './components/BlockContents'
-import { Container } from '../../components/UI/Layout'
+import HeaderAwareContainer from '../../components/UI/Layout/HeaderAwareContainer'
 
 export default class BlockScreen extends React.Component {
   render() {
     const { id, imageLocation } = this.props.navigation.state.params
 
     return (
-      <Container>
+      <HeaderAwareContainer>
         <BlockContents
           id={id}
           imageLocation={imageLocation}
         />
-      </Container>
+      </HeaderAwareContainer>
     )
   }
 }
