+++ screens/ChannelScreen/index.js
@@ -1,18 +1,18 @@
 import React, { Component } from 'react'
 
-import { Container } from '../../components/UI/Layout'
+import HeaderAwareContainer from '../../components/UI/Layout/HeaderAwareContainer'
 import ChannelContainer from './components/ChannelContainer'
 
 export default class ChannelScreen extends Component {
   render() {
     return (
-      <Container>
+      <HeaderAwareContainer>
         <ChannelContainer
           {...this.props}
           page={1}
           type="BLOCK"
         />
-      </Container>
+      </HeaderAwareContainer>
     )
   }
 }
