+++ screens/SearchScreen/index.js
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import SearchHeader from '../../components/SearchHeader'
 import TabToggle from '../../components/TabToggle'
 import SearchContents from '../../components/SearchContents'
-import { StatusBarAwareContainer } from '../../components/UI/Layout'
+import StatusBarAwareContainer from '../../components/UI/Layout/StatusBarAwareContainer'
 
 import navigationService from '../../utilities/navigationService'
